rename: 정보 관련 커스텀 훅 폴더 구조 변경
