+++ src/components/informations/edit/InformationUpdateEditor.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useInformationUpdateEditor } from "@/hooks/information/useInformationUpdateEditor";
+import { useInformationUpdateEditor } from "@/hooks/information/edit/useInformationUpdateEditor";
 import { InformationDetailDto } from "@/types/InformationDto";
 import { FormProvider } from "react-hook-form";
 import InformationEditor from "../common/InformationEditor";
