+++ src/hooks/diary/edit/useDiaryUpdateEditor.ts
@@ -2,8 +2,7 @@
 
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import usePreventBodyScroll from "../usePreventBodyScroll";
-import useModalBackHandler from "../useModalBackHandler";
+
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { DiaryUpdateFormSchema } from "@/lib/zod/schema/DiaryUpdateFormSchema";
@@ -13,6 +12,8 @@ import sanitizeHtml from "sanitize-html";
 import sanitizeOption from "@/constants/common/sanitizeOption";
 import { FEELING_STATUS } from "@/constants/diary/feelingStatus";
 import { fetchWithAuth } from "@/utils/fetchWithAuth";
+import usePreventBodyScroll from "@/hooks/usePreventBodyScroll";
+import useModalBackHandler from "@/hooks/useModalBackHandler";
 
 export const useDiaryUpdateEditor = (diaryData: GetDiaryResponseDto) => {
   const router = useRouter();
