+++ src/hooks/information/common/useCategoryModal.ts
