+++ src/hooks/information/detail/useImageViewer.ts
