+++ src/hooks/information/detail/useInformationLikeCount.ts
